refactor: move FormattableDuration
