+++ cmd/cmd_renew.go
@@ -141,7 +141,7 @@ func renewForDomains(ctx context.Context, cmd *cli.Command, lazyClient lzSetUp, 
 	// This is just meant to be informal for the user.
 	log.Info("acme: Trying renewal.",
 		log.CertNameAttr(certID),
-		slog.Any("time-remaining", FormattableDuration(cert.NotAfter.Sub(time.Now().UTC()))),
+		slog.Any("time-remaining", log.FormattableDuration(cert.NotAfter.Sub(time.Now().UTC()))),
 	)
 
 	err = hookManager.Pre(ctx, certID, renewalDomains)
@@ -230,7 +230,7 @@ func renewForCSR(ctx context.Context, cmd *cli.Command, lazyClient lzSetUp, cert
 	// This is just meant to be informal for the user.
 	log.Info("acme: Trying renewal.",
 		log.CertNameAttr(certID),
-		slog.Any("time-remaining", FormattableDuration(cert.NotAfter.Sub(time.Now().UTC()))),
+		slog.Any("time-remaining", log.FormattableDuration(cert.NotAfter.Sub(time.Now().UTC()))),
 	)
 
 	err = hookManager.Pre(ctx, certID, certcrypto.ExtractDomainsCSR(csr))
@@ -286,7 +286,7 @@ func isInRenewalPeriod(cert *x509.Certificate, certID string, days int, now time
 	log.Infof(
 		log.LazySprintf("Skip renewal: The certificate expires at %s, the renewal can be performed in %s.",
 			cert.NotAfter.Format(time.RFC3339),
-			FormattableDuration(dueDate.Sub(now)),
+			log.FormattableDuration(dueDate.Sub(now)),
 		),
 		log.CertNameAttr(certID),
 	)
@@ -433,44 +433,4 @@ func sameDomains(a, b []string) bool {
 	sort.Strings(bClone)
 
 	return slices.Equal(aClone, bClone)
-}
-
-type FormattableDuration time.Duration
-
-func (f FormattableDuration) String() string {
-	d := time.Duration(f)
-
-	days := int(math.Trunc(d.Hours() / 24))
-	hours := int(d.Hours()) % 24
-	minutes := int(d.Minutes()) % 60
-	seconds := int(d.Seconds()) % 60
-	ns := int(d.Nanoseconds()) % int(time.Second)
-
-	s := new(strings.Builder)
-
-	if days > 0 {
-		_, _ = fmt.Fprintf(s, "%dd", days)
-	}
-
-	if hours > 0 {
-		_, _ = fmt.Fprintf(s, "%dh", hours)
-	}
-
-	if minutes > 0 {
-		_, _ = fmt.Fprintf(s, "%dm", minutes)
-	}
-
-	if seconds > 0 {
-		_, _ = fmt.Fprintf(s, "%ds", seconds)
-	}
-
-	if ns > 0 {
-		_, _ = fmt.Fprintf(s, "%dns", ns)
-	}
-
-	return s.String()
-}
-
-func (f FormattableDuration) LogValue() slog.Value {
-	return slog.StringValue(f.String())
 }
